moved tiling logic to soilDrainage
