+++ src/pages/Location/SoilCondition/SoilDrainage/SoilDrainage.js
@@ -1,7 +1,9 @@
+/* eslint-disable max-len */
+
 import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { Button, Typography } from '@mui/material';
-import { LocalDrinkOutlined } from '@mui/icons-material';
+import { Button, Typography, Switch } from '@mui/material';
+import { LocalDrinkOutlined, InvertColors } from '@mui/icons-material';
 import { ReferenceTooltip } from '../../../../shared/constants';
 import arrayEquals from '../../../../shared/functions';
 import '../../../../styles/soilConditions.scss';
@@ -9,7 +11,7 @@ import RenderDrainageClasses from './RenderDrainageClasses';
 import MyCoverCropReset from '../../../../components/MyCoverCropReset/MyCoverCropReset';
 import { updateDrainageClass as updateDrainageClassRedux } from '../../../../reduxStore/soilSlice';
 
-const SoilDrainage = ({ setTilingCheck }) => {
+const SoilDrainage = () => {
   const dispatchRedux = useDispatch();
 
   // redux vars
@@ -19,7 +21,9 @@ const SoilDrainage = ({ setTilingCheck }) => {
   const myCoverCropListLocationRedux = useSelector((stateRedux) => stateRedux.sharedData.myCoverCropListLocation);
 
   // useState vars
+  const [showTiling, setShowTiling] = useState(false);
   const [handleConfirm, setHandleConfirm] = useState(false);
+  const [tilingCheck, setTilingCheck] = useState(false);
 
   useEffect(() => {
     if (myCoverCropListLocationRedux !== 'selector' && selectedCropsRedux.length > 0) {
@@ -27,6 +31,14 @@ const SoilDrainage = ({ setTilingCheck }) => {
       setHandleConfirm(true);
     }
   }, [selectedCropsRedux, myCoverCropListLocationRedux]);
+
+  useEffect(() => {
+    const checkArray = ['Very poorly drained', 'Poorly drained', 'Somewhat poorly drained'];
+    if (checkArray.some((e) => soilDataRedux?.drainageClass.includes(e))) {
+      setShowTiling(true);
+    }
+    window.localStorage.setItem('drainage', JSON.stringify(soilDataRedux?.drainageClass));
+  }, [soilDataRedux?.drainageClass]);
 
   const resetDrainageClasses = () => {
     dispatchRedux(updateDrainageClassRedux(soilDataOriginalRedux?.drainageClass));
@@ -99,6 +111,88 @@ const SoilDrainage = ({ setTilingCheck }) => {
         <RenderDrainageClasses drainage={soilDataRedux?.drainageClass} />
       </div>
       <MyCoverCropReset handleConfirm={handleConfirm} setHandleConfirm={setHandleConfirm} />
+      {showTiling && (
+        <div className="col-12 pt-2 mt-2 row">
+          <div className="col-12">
+            <Typography variant="body1" className="soilConditionSubHeader">
+              <InvertColors />
+              &nbsp;TILING &nbsp;
+              <ReferenceTooltip
+                type="text"
+                content="Indicate if the field of interest has tile installed. If you have selected very poorly to somewhat poorly drained soils, selecting “yes” will increase your drainage class by one factor."
+              />
+            </Typography>
+          </div>
+          <div className="col-12 pt-2">
+            <div className="pl-1 text-left">
+              <Typography variant="body1" display="inline">
+                NO
+              </Typography>
+              <Switch
+                checked={tilingCheck}
+                onChange={(e) => {
+                  const soilDrainCopy = soilDataRedux?.drainageClass;
+
+                  const drainSet = new Set(soilDrainCopy);
+                  if (e.target.checked) {
+                    if (
+                      drainSet.has('Very poorly drained')
+                      && drainSet.has('Poorly drained')
+                      && drainSet.has('Somewhat poorly drained')
+                    ) {
+                      drainSet.delete('Very poorly drained');
+                      drainSet.add('Moderately well drained');
+                    } else if (drainSet.has('Very poorly drained') && drainSet.has('Poorly drained')) {
+                      drainSet.delete('Very poorly drained');
+                      drainSet.add('Somewhat poorly drained');
+                    } else if (
+                      drainSet.has('Poorly drained')
+                        && drainSet.has('Somewhat poorly drained')
+                    ) {
+                      drainSet.delete('Poorly drained');
+                      drainSet.add('Moderately well drained');
+                    } else if (
+                      drainSet.has('Very poorly drained')
+                        && drainSet.has('Somewhat poorly drained')
+                    ) {
+                      drainSet.delete('Very poorly drained');
+                      drainSet.delete('Somewhat poorly drained');
+                      drainSet.add('Poorly drained');
+                      drainSet.add('Moderately well drained');
+                    } else if (drainSet.has('Very poorly drained')) {
+                      drainSet.delete('Very poorly drained');
+                      drainSet.add('Poorly drained');
+                    } else if (drainSet.has('Poorly drained')) {
+                      drainSet.delete('Poorly drained');
+                      drainSet.add('Somewhat poorly drained');
+                    } else if (drainSet.has('Somewhat poorly drained')) {
+                      drainSet.delete('Somewhat poorly drained');
+                      drainSet.add('Moderately well drained');
+                    } else {
+                      drainSet.delete('Very poorly drained');
+                      drainSet.delete('Poorly drained');
+                      drainSet.delete('Somewhat poorly drained');
+                      drainSet.add('Moderately well drained');
+                    }
+                    window.localStorage.setItem('drainage', JSON.stringify([...drainSet]));
+                  } else {
+                    window.localStorage.setItem(
+                      'drainage',
+                      JSON.stringify(soilDataOriginalRedux?.drainageClass),
+                    );
+                  }
+
+                  setTilingCheck(!tilingCheck);
+                }}
+                name="checkedC"
+              />
+              <Typography variant="body1" display="inline">
+                YES
+              </Typography>
+            </div>
+          </div>
+        </div>
+      )}
     </div>
   );
 };
